+++ usuarios_app/urls.py
@@ -1,8 +1,7 @@
 from django.contrib.auth import views as auth_views
 from django.urls import path
 
-from .views import CriarUsuario
-from carrinho_app import shopping_items_add
+from .views import CriarUsuario, shopping_items_add
 
 login_view = auth_views.LoginView.as_view(
     template_name="usuarios_app/login.html",
